Renamed 10_ball_sr to resolve import issues
